chore: Create new folder to contain project in solution
