+++ ClassiVault.Api.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClassiVault.Api", "ClassiVault.Api.csproj", "{5515F605-C74A-4081-8B66-F0CF0AC99F34}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClassiVault.Api", "ClassiVault.Api/ClassiVault.Api.csproj", "{5515F605-C74A-4081-8B66-F0CF0AC99F34}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
