+++ ClassiVault.Api/.gitignore
