+++ ClassiVault.Api/Controllers/HealthCheckController.cs
