+++ ClassiVault.Api/DataAccess/Models/EncryptionKeyInfo.cs
