+++ ClassiVault.Api/DataAccess/Models/User.cs
