+++ ClassiVault.Api/Services/IEncryptionKeyInfoService.cs
