+++ ClassiVault.Api/Services/PasswordVaultService.cs
