--- Services/IPasswordVaultService.cs
@@ -1,10 +1,0 @@
-using ClassiVault.Api.DataAccess.Models;
-
-public interface IPasswordVaultService
-{
-  Task<List<PasswordVault>> GetAllAsync();
-  Task<PasswordVault?> GetOneAsync(long id);
-  Task<PasswordVault> AddAsync(PasswordVault passwordVault);
-  Task<PasswordVault> UpdateAsync(PasswordVault passwordVault);
-  Task DeleteAsync(long id);
-}
